BUG#13389854 - VALGRIND ERROR IN _MI_WRITE_BLOB_RECORD
When a record is inserted into a MyISAM temporary table using
fixed record format, the entire content of the record buffer is
written to the table. If one or more of the columns is a 
VARCHAR, the trailing space may be uninitialized.

In this bug, trailing and unitialized space from a VARCHAR was 
copied from an InnoDB fulltext index and stored in a MyISAM 
temp table. This made valgrind bark.

The fix is to restrict copying of VARCHARs in field_conv() to
only the bytes actually used. This is implemented by:
  * refactoring: merge do_varstring*() into a single function.
  * refactoring: make do_varstring() use a utility function 
    copy_field_varstring().
  * bugfix: make field_cond() use utility function 
    copy_field_varstring() if the two input fields are 
    varchars with the same number of length bytes.