BUG#11756433 (MySQL48348): ADD TRANSLATION OF ERROR CODE TO ERROR MESSAGE IN ERROR LOG
A lot of our error messages were of the type
"failed to XYZ because of error 123",
essentially requiring the DBA to look up error 123
with the perror tool.

If it's worth printing, it's worth printing right,
so we now do
"Can't get stat of '/tmp/foo' (Errror 2, no such file/dir"
This works for both OS errors and handler errors as the
latter do come up as reasons for higher order errors
("can't rename table: 150 - foreign key constraints apply").