BUG#12844302: 62055 RACE CONDITION IN CHECK_TEMP_DIR() FROM MULTIPLE MYSQLD INSTANCES
Problem:
=========
In check_temp_dir(), mysqld creates a temporary file and removes  it immediately, 
without protecting any mutex/lockfile/etc. A problem might happen when multiple
mysqld instances on the same host start SQL thread at the same time. SQL thread
might abort with an error.

Background:
=========
The temp files being created when the SQL thread is started, and during the LOAD
DATA INFILE operation does not guarantee to prevent any collision in the filename
when multiple slaves are started on a single machine(host), that share the same
temp directory.

Resolution:
=========
We resolve this problem by adding the server UUID to temp file created by
check_temp_dir(). This guarantees that the files created in the temp directory have
unique names. We also added server UUIDs of the slave in the temp file created so
that there will be no collision in the file name when multiple slaves share the temp
directory. e.g. during LOAD DATA INFILE Query.