Bug#12953735 - INNODB; DDL CHECK FOR UNDO RECORD TOO BIG IS NOT CORRECT
This patch does not delete the incorrect code, but instead puts
 #if 1 … #endif
around it with a comment that it should be deleted once UNDO records
can spill over from one page to the next.

It also contains enhancements to innodb_16k.test, innodb_8k.test,
innodb_4k.test and innodb_prefix_index_liftedlimit.test that show how
the current existing code in dict_index_too_big_for_undo() is not
accurate.  The three page specific tests show how a table can be created
and a record inserted that CANNOT be updated because the resulting UNDO
record is too big.  The other test, innodb_prefix_index_liftedlimit.test,
has a change that indicates where the function rejects an index, but if
it did not, the record would be OK to update.