Bug#13259227 ASSERTION TABLES->TABLE->POS_IN_TABLE_LIST              == TABLES FAILED IN SQL_BASE.CC:4668
This assert was triggered if INSERT/REPLACE DELAYED was used
with a temporary table.

The problem was that temporary tables are pre-opened before
INSERT/REPLACE starts even if INSERT/REPLACE DELAYED does not
support temporary tables. This meant that when
Delayed_insert::open_and_lock_table() was later executed, 
TABLE_LIST::table was already pointing to the temporary table.
This eventually caused the assert to be triggered.

This patch fixes the problem by not pre-opening temporary
tables for INSERT/REPLACE and instead opening them only when
we know that it's not a DELAYED insert.

Note that the restriction that temporary tables are not supported
for INSERT/REPLACE DELAYED should be reflected in the manual.
(This is not new behavior, it's the same for 5.1/5.5.)

This bug was a regression caused by the patch for                         
Bug#11746602.

Test case added to delayed.test.