Bug#13414014: Extra rows in result on semijoin query with where +               non-matching where predicate
The problem concerns a query using semi-join MaterializeLookup,
a multi-level combined condition after semi-join conversion, and
equality propagation.
Say we have two tables t1(a, b, c, d) and t2(a) and the query:

SELECT * FROM t2 AS ot
WHERE (a, a) IN (SELECT a, b FROM t1 AS it
                 WHERE it.a = 'x' OR it.c > it.d);

This is transformed to a semi-join, and we end up with this condition
attached to the query object:

( =('x', ot.a, it.a, it.b) or (it.c > it.d)) and =(ot.a, it.a, it.b)

=(...) is shorthand for a multiple equality covering the expressions
within the parentheses.

With MaterializeLookup strategy, substitute_for_best_equal_field()
will transform the above condition to this:

(it.a = ot.a) and (it.b = it.a) and ((ot.a = 'x') or (it.c > it.d))

Notice here that we have transformed the original predicate (it.a = 'x')
to (ot.a = 'x'). In a regular join this makes sense, because ot comes
before it in join order, but in a materialization, we have no access
to ot while materializing it.

This is handled properly when the predicate is not inside an OR
condition. If the inner condition was simply it.a = 'x', we would
have these conditions
before substitution: =('x', it.a, ot.a, it.b)
after  substitution: ot.a = 'x' and it.a = 'x' and it.b = 'x'.

Back to the problem at hand, make_join_select() generates
table conditions:

Table condition for ot: (nil)
Table condition for it: (it.b = it.a)

(it.a = ot.a is not present because it is handled by MaterializeLookup).
As we see, the check on literal 'x' is gone, and so is the less than
predicate. This happens because we generate table conditions for either
the table set [ot] or [it], never for [it, ot] combined, which is
actually needed for the join expression. The less than predicate is
not generated because it is part of an OR condition, and the rest of
the OR condition are not generated, and thus assumed to be TRUE.

The invalid transformation to (ot.a = 'x') happens inside
eliminate_item_equal(). There are two multiple equalities here,
the one containing the literal value 'x' on the lower level, and the
other at the upper level. When adding equalities for a multiple
equality at a lower level, the function makes sure that the
equality is not already covered at an upper level. Unfortunately,
this requires that all tables are available for condition processing,
which is not necessarily true when dealing with a materialized
semi-join nest.

Hence, to make sure that all necessary predicates are available during
materialization, we make sure that equalities for materialized tables
are generated, even though they are covered by upper-level multiple
equalities.

After the bugfix, the transformed condition is:

(it.a = ot.a) and (it.b = it.a) and
(((ot.a = 'x') and (it.a = 'x') and (it.b = 'x')) or (it.c > it.d))

The table condition generated by make_join_select() for table it is:

it: (it.b = it.a) and (((it.a = 'x') and (it.b = 'x')) or (it.c > it.d))

mysql-test/include/subquery_sj.inc
  Added test case for bug#13414014.

mysql-test/r/subquery_sj_all.result
mysql-test/r/subquery_sj_all_bka.result
mysql-test/r/subquery_sj_all_bka_nixbnl.result
mysql-test/r/subquery_sj_all_bkaunique.result
mysql-test/r/subquery_sj_dupsweed.result
mysql-test/r/subquery_sj_dupsweed_bka.result
mysql-test/r/subquery_sj_dupsweed_bka_nixbnl.result
mysql-test/r/subquery_sj_dupsweed_bkaunique.result
mysql-test/r/subquery_sj_firstmatch.result
mysql-test/r/subquery_sj_firstmatch_bka.result
mysql-test/r/subquery_sj_firstmatch_bka_nixbnl.result
mysql-test/r/subquery_sj_firstmatch_bkaunique.result
mysql-test/r/subquery_sj_loosescan.result
mysql-test/r/subquery_sj_loosescan_bka.result
mysql-test/r/subquery_sj_loosescan_bka_nixbnl.result
mysql-test/r/subquery_sj_loosescan_bkaunique.result
mysql-test/r/subquery_sj_mat.result
mysql-test/r/subquery_sj_mat_bka.result
mysql-test/r/subquery_sj_mat_bka_nixbnl.result
mysql-test/r/subquery_sj_mat_bkaunique.result
mysql-test/r/subquery_sj_mat_nosj.result
mysql-test/r/subquery_sj_none.result
mysql-test/r/subquery_sj_none_bka.result
mysql-test/r/subquery_sj_none_bka_nixbnl.result
mysql-test/r/subquery_sj_none_bkaunique.result
  Added test case results for bug#13414014.

sql/sql_select.cc
  In eliminate_item_equal(), when dealing with expressions from tables
  used in materialized semi-join nests used for MaterializeLookup,
  make sure that predicates generated from multiple equalities, that
  are normally not necessary because they are covered by an upper level
  multiple equality, are still generated.