A patch for Bug#11748352 - 36002: PREPARED STATEMENTS: IF A VIEW USED IN A STATEMENT IS REPLACED, BAD DATA.
The problem was that changes in view meta-data were not
detected by prepared statements.

There were the following technical causes for this problem:

  - TABLE_SHARE ref-version was always set to 0 for views;

  - ALTER VIEW did not flush corresponding TABLE_SHARE-object
    from the table definition cache.

The patch has the following fixes:

  - use TABLE_SHARE::table_map_id as ref-version for views
    (like it is done for regular tables).

  - explicitly flush TABLE_SHARE-object of the view after
    ALTER VIEW statement.

  - check "table version" when opening a view under LOCK TABLES.

The test case for Bug#36002 was already in ps_ddl.test, so no need
for adding a new test case. The results however were corrected.
Also, tests for altering views and handling changes in the locked
table mode were added.

Note, that this patch introduces a minor backward-incompatible change:
instead of successfully returning wrong data, the server now may throw
a table-not-found error. It happens in cases like the following:
  - create a base table;
  - create a trigger on that table (let's say after-insert-trigger);
  - access a view (do SELECT or a DML-statement) inside that trigger;
  - prepare a statement that inserts data into the base table;
  - execute the prepared statement at least once;
  - re-create the view, or alter it so that it uses a different set of
    base tables;
  - execute the prepared statement.
    This execution used to lead to wrong data, because the outdated
    view definition was used. Now, it leads to table-not-found error.
    It's expected that such scenarios will be fixed by WL#4179.

So, to generalize the scenario:
  - there need to be a stored program, using a view;
  - there need to be a prepared statement causing the stored program
    invocation;
  - if the view gets altered (or re-created) in the way that it
    uses a different set of base tables, further executions of
    the prepare statements result in a table-not-found error.

A workaround is to use FLUSH TABLES after meta-data changes
(the workaround is not affected by the patch).