Bug#12773464 - 61925: WRONG RESULT WITH AGGREGATE +                NOT BETWEEN + KEY
The problem: opt_sum_query() was invoked to replace MAX(a) with
a constant in the query:
   SELECT MAX(a) FROM t1 WHERE a NOT BETWEEN 3 AND 9;

However, opt_sum_query() mistook NOT BETWEEN for a BETWEEN
and therefore found the highest values lower than or equal to 9
and then tested the predicate on this value. The predicate 
obviously evaluated to false and MAX(a) therefore evaluated to
NULL.

opt_sum_query() can only handle conjunctions ("AND" predicates)
but "x NOT BETWEEN a AND b" is equivalent to "x < a OR x > b" 
and is therefore not a conjunction. The fix is to return
from the function without doing const replacement for 
NOT BETWEENs.