ndb - new template for intrusive list
The M in DLMList and SLMList can mean Meta or Methods, since the
template takes a Meta/Methods-class as parameter to tell the list
how to get next/prev link.  This to be able to use the list structures
with classes that names nextList/prevList in another way.

The default Meta/method-class assumes the same interface that is already
used in classes using DLList and SLList.

DLMList and SLMList logic are copied from DLListImpl and SLListImpl
respectively.  Direct references like obj->nextList are replaced with
M::nextList(*obj) and the like for prevLike.