BUG#13350136 - TEMPORARY MYISAM TABLES DO NOT PACK LONG                VARCHAR COLUMNS
MyISAM has three record formats (fixed, packed and compressed) 
out of which fixed and packed are applicable for temporary 
tables. The difference is that with fixed length record format,
VARCHAR trailing bytes are written whereas for packed record 
format the trailing bytes are not written. The packed record 
therefore saves space at the cost of some additional memcopies.

The packed record format is currently not in use by temporary
tables, although old commit messages suggests that this has not
always been the case, e.g:

   revision: sp1r-monty@mysql.com-20041206000037-04661
   Date: 2004-12-06
   "Internal temporary files can now use fixed length tables 
    if the used VARCHAR columns are short"

This CS enables packed record format for temporary tables.

The CS also includes a fix for a MyISAM bug that is only a
problem when the packed format and a special unique constraint
(that checksums wide columns) is used. This unique constraint
only applies to internal temporary tables when uniqueness is
required on columns with a total size bigger than normal 
indexes allow (1000 bytes). The problem:

A column's value in MyISAM can be read from or written to the 
row buffer using MI_COLUMNDEF::offset ("normal" column 
definition) or HA_KEYSEG::start (key definition). The former 
is calculated in mi_open() and is the sum of the length of all
preceding columns. The latter is calculated in mi_create(). 
The bug was that in mi_create(), HA_KEYSEG::start was set to 
the offset of the column for the on-disk format (reclength) 
whereas it should have been the offset for the in-memory format.

Due to this bug, unique constraint values were written to one
offset (HA_KEYSEG::start in mi_check_unique()) and read from
another offset (MI_COLUMNDEF::offset in _mi_pack_record et al),
resulting in duplicate rows in the result set because equally
valued rows got seemingly different checksums.