BUG#11751736 - 42708: DROP DATABASE STATEMENT SHOULD REMOVE .OLD SUFFIX FROM 		      DATABASE DIRECTORY
BACKGROUND:
Create database, tables and load data.Execute myisampack to
compress the tables, also use -b option to have backup of 
datafile.Try dropping the database, will fail because of 
.OLD file that is created in database directory.
	
FIX:
DROP DATABASE will fail with the error described in the bug report
if an unknown file is found while trying to delete the database
directory. This bug is fixed by adding ".OLD" to the list of known file
extensions that DROP DATABASE uses when checking if a file can be
silently deleted.