Bug#12911480: GROUP BY MAKE SET WITH EMPTY SET VALUES CRASH IN MY_HASH_SORT_SIMPLE
Problem Description:
While executing a make_set funtion in a group by clause with null output,
 we see problems like, accessing uninitialized area.

Problem Analysis:
While deducing the nullability of make_set function 
(in Item_func_make_set::fix_fields), we do not take into consideration that the 
first argument can be null and as a result the overall function result could be 
null. 

In the scenario presented, the first argument is null. So the result of 
make_set will be null and therefor while accessing the keys in "hp_hashnr" 
function, it tries to access an uninitialized area (as null bit is not set) 
and hence the valgrind error. 

While we do consider that the result set can be null when one of the arguments 
(starting from the second) can be null in a make_set function. 

In the current scenario, which is "select count(*) from t1 group by make_set(a,b)", 
we do not consider field "a"s nullability. We only consider the rest of the 
argument list which is "b" here.  As "b" is defined as not null while creating the 
table, we are seeing the problem. 

Solution:
In Item_func_make_set::fix_fields(), we take into consideration even the first 
argument's nullability.