Bug#11764371 57196: MORE FUN WITH ASSERTION: !TABLE->FILE || TABLE->FILE->INITED == HANDLER::
Problem: An assertion is triggered by queries that assign the result
of a subquery to a user variable when the outer query is doing
DISTINCT and GROUP BY.

Because of the DISTINCT and GROUP BY, the query is executed using two
temporary tables. The first temporary table is filled by executing the
query over the base tables. The subquery is also executed, and the
result is stored in the user variable. After this execution,
join_free() is called to free data from the old join and release all
read locks since the next execution is based on the temporary table.

The second round of execution is over the first temporary table, but
because of the user variable assignment, the subquery is re-executed
over the base tables. Since these tables are now unlocked, an
assertion guarding handler::ha_rnd_next() against reading from
unlocked tables is triggered.

Fix: In change_to_use_tmp_field(), if the item is an
Item_func_set_user_var that stores the value of a subquery, replace it
with a cloned Item_func_set_user_var that stores the result of an
Item_field that contains the result of the subquery. This makes the
next execution round use the stored value instead of re-executing the
subquery.

This patch also fixes bug #13260504. No additional testcase required.