Revert sunny.bains@oracle.com-20111103014755-9zzp2oscgvem35th
The asm code that was introduced in that revision causes a crash in
ut_cpuid() when that function is not inlined, e.g. compiled with an
attribute ``__attribute__((noinline))'' or with gcc's -fno-inline.
The crash only happens on some 64 bit platforms, fimafeng09.no.oracle.com
for example.

It is not clear what sunny.bains@oracle.com-20111103014755-9zzp2oscgvem35th
was trying to address, but the old code (before that changeset) does pass all
tests currently.

Should any problems arise later (the ones that the reverted changeset was
addressing) and if they are due to C vs. C++ difference we could add
``extern "C"'' to the function ut_cpuid().

Discussed with:	Marko and Sunny (via IM)