Patch for bug#12872824 (formerly known as 62125): testing stored function result for null incorrectly yields 1292 warning.
The reason for this bug was that when server executed expression 'IS NULL'
for return value of some stored function it calculated the return
value before checking it for NULL. During return value calculation
it's assumed that a value has INTEGER type that is not true in general.

The solution is to evaluate return value for NULL without its calculation.