WL#5534 Online ALTER, Phase 1
This worklog extends the handler API to add general support
for online table changes and rewrites the ALTER TABLE
implementation to it. The handler API extensions are based
on Cluster's online ALTER implementation and therefore also
narrows the gap between the Server and Cluster codebases.

Using this new API, the server supports concurrent reads
and writes during the main phase of in-place ALTER TABLE
operations. In this first phase the focus is on index
creation, but the API is general enough to be extended to
other ALTER TABLE operations.

The ALTER TABLE syntax is extended to allow users to specify
concurrency requirements as well as select between the in-place
and copy algorithms.

Execution of an ALTER TABLE statement has three phases:
1) Initialization
   In this phase the algorithm and locking level is determined.
   This is done based on requirements from the user, capabilities
   of the given storage engine and characteristics of the
   ALTER TABLE operation.
   The new handler API calls for this phase has a default
   implementation that converts them to the old API calls so
   existing engines can be used without changes.

2) Execution
   In this phase, the table definition is changed. This is done
   either directly on the current table (i.e. in-place) or using
   a temporary table copy. The new API calls only concern the
   the former. In this patch, a compatibility layer converts the
   new API calls for in-place execution to the old API calls for
   index creation/deletion. However these implementations will be
   removed later once the InnoDB changes that makes use of the
   new API is in place.

3) Finalization
   The final phase, updates the data dictionary, installs the
   new .FRM and notifies the storage engine.

The metadata subsystem is extended with a new lock type,
MDL_SHARED_UPGRADABLE, which is used for phases when ALTER TABLE
allows concurrent reads and writes.

The patch also updates the partitioning implementation to use the
new handler API functions.