Fix Bug#13513676 INNODB_BUFFER_POOL_DUMP_STATUS IS NOT POPULATED WITHIN EXPECTED TIME
There is a tiny interval of time during startup when the status variable
'innodb_buffer_pool_dump_status' can be '' before it becomes 'not started'.

This patch makes the bug 100% reproducable:

--- cut ---
  --- storage/innobase/buf/buf0dump.cc
  +++ storage/innobase/buf/buf0dump.cc
  @@ -577,12 +577,13 @@ DECLARE_THREAD(buf_dump_thread)(
   /*============================*/
   	void*	arg __attribute__((unused)))	/*!< in: a dummy parameter
   						required by os_thread_create */
   {
   	srv_buf_dump_thread_active = TRUE;
   
  +	sleep(5);
   	buf_dump_status(STATUS_INFO, "not started");
   	buf_load_status(STATUS_INFO, "not started");
--- cut ---

It is not easy to assign a default value to this variable which to be
available before buf_dump_thread() starts, thus:

The solution is to use include/wait_for_status_var.inc to wait for the
variable to become 'not started'.