Post-push fix for BUG#13728026.
After pushing the patch for BUG#13728026, some tests started failing
on valgrind because the repository was not being registered in the
Relay_log_info object. So when the thread and the Relay_log_info were
destroyed the memory allocated for the repository was not freed
causing a memory leak.

The Relay_log_info object is responsible for destroying the repository.

In this patch, we correctly register the repository in Relay_log_info
object.