Fix for WL#5558: Resolve ORDER BY execution method at the optimization stage
Recommit, including review comments, and some new tests after
trunk-backporting was merged into trunk.

The decision whether to sort result dataset or use ordered index scan to get 
sorted result was partly made in JOIN::optimize, and partly in JOIN::execute().
Furthermore the logic for EXPLAIN'ing the sort was different/incorrect compared
to the logic used for determine filesort or not during execution. This sometimes
caused EXPLAIN to be incorrect.
      
This refactoring WL moves the decision on how to handle ORDER BY to the 
optimization stage. It is a pure refactoring in the sense that it
does not change the current filesort cost model. However, several 
test results has changed as the EXPLAIN of those used to be incorrect.
During the refactoring task several other bugs has also been 
reported and fixed as standalone bugs, these are: 
      
 - bug 13514959: query plan fails to 'using index' where it should
 - bug 13528826: test_if_cheaper_ordering(): calculates incorrect 'select_limit'
 - bug 13529048: test_if_skip_sort_order() don't have to filesort a single row.
 - bug 13531865: test_if_skip_sort_order() incorrectly skip filesort if 'type' is ref_or_null

As of today these fixes has already been reviewed and pushed into mysql-trrunk.
      
As an extension of the refactoring in this commit, one should considder
to move test_if_skip_sort_order() w/ family from sql_select.cc 
into sql_optimize.cc as this is new the only place it is used.
(make it 'static' also)
      
There is likely also some code in JOIN::execute() related to
maintaining 'pre_idx_push_cond' which now is obsolete and
should be considdered for removal. (later)