BUG#11762377 - 54963: ENHANCE THE ERROR MESSAGE TO REDUCE USER CONFUSION
BACKGROUND:
If a user tries to create a table with columns that combined
have a total length larger than the maximum row size, an error
message is returned. In this error message, the maximum row 
length of the storage engine is given. The problem is that 
it is not possible to create a table with a column length 
matching the given maximum row length, as there are additional 
column storage overheads that the error message does not 
account for.

FIX:
This bug is fixed by enhancing the error message mentioning
that the maximum size for columns is including storage 
overhead and the user is asked to consult the manual.A test
case has been written in create.test and the corresponding 
result file has also been updated.