BUG#11764155 - INCONSISTENT 'OUT OF RANGE' WARNING MESSAGE ON                SELECT
When the optimizer decides to use ref access on a table, the
conditions that are guaranteed to be satisfied by the ref 
access itself does not need to be checked in the server.

Removal of such "guaranteed by ref access" conditions is done
by make_cond_for_table_from_pred(). One way a predicate can be
guaranteed to be satisfied is if the consition is on the
form "field = constant" and the field is covered by the index.
In addition, it is required that the constant can be converted 
to the field's type without error when test_if_ref() saves the 
const value in the field.

Consider the following query where the constant is higher than
the maximum int value:

SELECT * FROM t1 WHERE int_col = 999999999999

The problem in this bug was that the value was stored using
thd->count_cuted_fields == CHECK_FIELD_WARN. Because of this,
store_val_in_field() would emit a warning when the too-high
integer in the above query was stored. There is no use for this
warning. In fact, when the range optimizer does a similar thing
when building the ranges, get_mm_leaf() calls
Item::save_in_field_no_warnings() instead to turn off warnings. 

The fix is a simplification: the function store_val_in_field() 
is removed and replaced with Item::save_in_field_no_warnings().
store_val_in_field() was used only to store constants in key
fields and was almost identical to
Item::save_in_field_no_warnings(). The difference between these
functions was that 
  a) the latter did not give warnings (which is desirable), and
  b) the latter accepts storing a 0000-00-00 date even when 
     the SQL mode does not permit it. This is OK and even 
     desirable because store_val_in_field() was only called for
     queries. Before this change, a query of the form 

        "SELECT * FROM t1 WHERE date_col = '0000-00-00'" 

     that was ref-accessed through an index covering date_col, 
     would not only evaluate the condition by ref-accessing in
     the index but also in the server because 
     store_val_in_field() reported "hey! NULL dates are not 
     allowed".