Bug#13736664 RIGHT JOIN+WHERE QUERY GIVES DIFF RESULTS WHEN USING SELECT * SUBQ FOR TABLES
Consider the query from the test case:

SELECT d
FROM (SELECT * FROM t2) AS a2 RIGHT JOIN (SELECT * FROM t1) AS a1 ON c = a
WHERE d LIKE '_'
ORDER BY d;

The query is executed by reading t1 as a const table and calling
create_sort_index() to read and sort t2 on column d. The join and
WHERE conditions are checked during filesort().

First, the optimizer decides to read a2 using the JT_REF access
method, so the join condition is removed and replaced with a ref
lookup. Later, when the optimizer decides to first do the filesort and
then the join, the condition is reintroduced since filesort() always
does a full table scan. The condition is added by
add_ref_to_table_cond(), which calls create_cond_for_const_ref() to
create the condition. An Item_func_equal is constructed to compare c
and a. This is equivalent to checking c <=> a, so if a and c are both
NULL, the rows match the join condition and are part of the
result. This is clearly not the same as the original query, so the
query returns rows that shouldn't match the join condition.

Fix: In create_cond_for_const_ref(), create an Item_func_eq instead of
Item_func_equal if join_tab->ref.null_rejecting is set for the key
part.