Bug#13803810: TOO FEW ROWS RETURNED FOR RANGE ACCESS               IN VARCHAR INDEX USING DATETIME VALUE
There are many allowed ways to represent DATE and TIME in MySQL,
e.g the default "2001-01-01 11:22:33" but also variations like
"01#01#2001 11,22,33" and "1.1.2001". If a VARCHAR column is 
used to store temporal data, we don't know which format is 
used and it may not even be consistent from row to row. Due to 
this, a lookup key cannot be created for temporal data into a 
VARCHAR index.

'ref' access of temporal values in VARCHAR columns was 
already disallowed, but 'range' access was not. The fix is to
also disallow 'range' access.