ndb - make DynArr256 release empty pages
DynArr256 use pages with 30 nodes each.
Each node can contain 256 words.

DynArr256Pool are used for seize and release
of nodes and it allocates new pages if necessary
and now also releases pages when no nodes on the
page are allocated.

The pool keeps a double linked list of pages with
at least one node free. Allocation of a node are
made from the first page in that list.

Within a page the first free node are used based
on its index.
The last free node are used to keep the links of
the double linked list of pages with free nodes.
Which nodes that are free are determined by bits
in the headers magic fields.